Don't GC on `count_to`
This commit changes `count_to` to use a while loop and to avoid touching
the GC.  `count_to` is not thread safe, so it was possible for threads
to context switch in the middle of the function, causing the GC to be
enabled when we didn't expect.  Better to just leave the GC alone.